Rename package to masker
